chore: move some messages internal
