+++ src/components/internals/messages/localized-input/index.js
@@ -1,0 +1,1 @@
+export { default } from './messages';
